+++ makefile
@@ -1,2 +1,2 @@
 main:chip8.cpp
-	g++ -o main chip8.cpp `sdl2-config --cflags --libs`
+	g++ -o main main.cpp `sdl2-config --cflags --libs`
